+++ src/main/java/roidrole/patternbanners/config/ConfigGeneral.java
@@ -3,6 +3,8 @@ package roidrole.patternbanners.config;
 
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.Configuration;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 import roidrole.patternbanners.Tags;
 
 import java.io.File;
@@ -28,9 +30,9 @@ public class ConfigGeneral {
         @Config.RequiresMcRestart
         public boolean craftable = true;
 
-        @Config.Comment("Should we generate pattern items for shapes?")
+        @Config.Comment("Should we generate pattern items for shapes?\nOnly affects /patternbanners:configupdate")
         @Config.RequiresMcRestart
-        public boolean shapes_pattern = true;
+        public boolean shapes_pattern = false;
 
         @Config.Comment("Set to true for the fallback texture to be the 1.14 texture")
         public boolean fallbackToOld = false;
@@ -42,6 +44,7 @@ public class ConfigGeneral {
     public static final RecipeCategory recipes = new RecipeCategory();
     public static class RecipeCategory{
 
+        @Config.Name("Pattern Apply")
         public final PatternApplyCategory patternApply = new PatternApplyCategory();
         public static class PatternApplyCategory{
             @Config.RequiresMcRestart
@@ -54,22 +57,25 @@ public class ConfigGeneral {
             @Config.RequiresMcRestart
             public boolean consumePattern = false;
 
-            @Config.Ignore
             public boolean enabled = loom || craftingTable;
         }
 
+        @Config.Name("Pattern Only Shape")
+        @Config.Comment("Concerns the banner patterns with no item nor pattern.\nReadds the vanilla crafting table recipe in crafting table, the scroll bar in loom")
         public final PatternOnlyShapeCategory patternOnlyShape = new PatternOnlyShapeCategory();
         public static class PatternOnlyShapeCategory{
             @Config.RequiresMcRestart
             public boolean loom = true;
 
-            //TODO:implement
             @Config.RequiresMcRestart
             public boolean craftingTable = false;
 
-            @Config.Ignore
             public boolean enabled = !ConfigGeneral.patterns.shapes_pattern && (loom || craftingTable);
         }
+
+        @SideOnly(Side.CLIENT)
+        @Config.Comment("Determines the layout of the HEI tab:\ntrue: Crafting Table\nfalse: Loom")
+        public boolean JEICraftingTable = patternApply.craftingTable || patternOnlyShape.craftingTable;
     }
     @Config.Ignore
     public static Configuration config = new Configuration(new File("config/"+Tags.MOD_ID+"/general.cfg"));